+++ commandsheet/cli.py
@@ -43,6 +43,11 @@ class CustomArgumentParser(argparse.ArgumentParser):
 def options(parser):
     opts = parser.add_argument_group(_('options'))
     opts.add_argument(
+        '--sample-config',
+        action='store_true',
+        help=_("Produce a sample `commandsheet.ini` file")
+    )
+    opts.add_argument(
         '-s',
         '--section-numbers',
         action='store_true',
@@ -78,7 +83,7 @@ def options(parser):
         '-V',
         '--version',
         action='version',
-        version='%(prog)s 0.1.0',
+        version='%(prog)s 0.2.0',
         help=_("Show version")
     )