+++ commandsheet/commandsheet.py
@@ -9,12 +9,14 @@ from commandsheet.config import parse_config
 from commandsheet.config import config_exists
 from commandsheet.config import is_valid_config_file
 from commandsheet.config import config_file_path_exists
+from commandsheet.config import produce_sample_config
 from commandsheet.compatibility import compatible_os
 from commandsheet.errors import not_a_valid_config_file
 from commandsheet.errors import no_config_file_found
 from commandsheet.errors import no_config_file_sections_found
 from commandsheet.errors import no_config_file_path_exists
 from commandsheet.errors import no_compatible_os
+from commandsheet.errors import too_many_fillchars
 from commandsheet.output import header
 from commandsheet.output import display_commandsheet
 
@@ -28,7 +30,14 @@ def main():
     if not compatible_os(operating_system):
         no_compatible_os(parser, operating_system)
 
+    sample_config = args.sample_config
+    if sample_config:
+        produce_sample_config(parser)
+
     fillchar = args.fillchar
+    if len(fillchar) > 1:
+        too_many_fillchars(parser)
+
     section_numbers = args.section_numbers
     config_file = args.config_file if args.config_file else XDG_CONFIG_PATH